address review comments and make node utils pkg
