+++ common/utils/net.go
@@ -11,13 +11,8 @@ import (
 	"crypto/x509"
 	"fmt"
 	"net"
-	"os"
-	"os/signal"
 	"strconv"
 	"strings"
-	"syscall"
-
-	"github.com/hyperledger/fabric-lib-go/common/flogging"
 
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/peer"
@@ -125,29 +120,4 @@ func TrimPortFromEndpoint(endpoint string) string {
 	}
 
 	return endpoint
-}
-
-type NodeStopper interface {
-	Stop()
-}
-
-func StopSignalListen(stopChan chan struct{}, node NodeStopper, logger *flogging.FabricLogger, nodeAddr string) {
-	signalChan := make(chan os.Signal, 1)
-	signal.Notify(signalChan, syscall.SIGTERM)
-
-	go func() {
-		defer signal.Stop(signalChan)
-
-		for {
-			select {
-			case <-signalChan:
-				logger.Infof("SIGTERM signal caught, the node listening on %s is about to shutdown:", nodeAddr)
-				node.Stop()
-				return
-			case <-stopChan:
-				logger.Infof("Exit StopSignalListen routine")
-				return
-			}
-		}
-	}()
 }
