+++ test/utils_test.go
@@ -37,7 +37,6 @@ import (
 	"github.com/hyperledger/fabric-x-orderer/common/types"
 	"github.com/hyperledger/fabric-x-orderer/config"
 	ordererRulesMocks "github.com/hyperledger/fabric-x-orderer/config/verify/mocks"
-	node2 "github.com/hyperledger/fabric-x-orderer/node"
 	"github.com/hyperledger/fabric-x-orderer/node/assembler"
 	"github.com/hyperledger/fabric-x-orderer/node/batcher"
 	"github.com/hyperledger/fabric-x-orderer/node/comm"
@@ -49,6 +48,7 @@ import (
 	"github.com/hyperledger/fabric-x-orderer/node/crypto"
 	protos "github.com/hyperledger/fabric-x-orderer/node/protos/comm"
 	"github.com/hyperledger/fabric-x-orderer/node/router"
+	node_utils "github.com/hyperledger/fabric-x-orderer/node/utils"
 	configMocks "github.com/hyperledger/fabric-x-orderer/test/mocks"
 	"github.com/hyperledger/fabric-x-orderer/testutil"
 	"github.com/hyperledger/fabric-x-orderer/testutil/client"
@@ -200,7 +200,7 @@ func createAssemblers(t *testing.T, num int, ca tlsgen.CA, shards []node_config.
 		logger := testutil.CreateLogger(t, i+1)
 		loggers = append(loggers, logger)
 
-		assemblerGRPC := node2.CreateGRPCAssembler(assemblerConf)
+		assemblerGRPC := node_utils.CreateGRPCAssembler(assemblerConf)
 
 		assembler := assembler.NewAssembler(assemblerConf, assemblerGRPC, genesisBlock, logger)
 		assemblers = append(assemblers, assembler)
@@ -518,7 +518,7 @@ func recoverConsenter(t *testing.T, ca tlsgen.CA, conf *node_config.ConsenterNod
 }
 
 func recoverAssembler(t *testing.T, conf *node_config.AssemblerNodeConfig, logger *flogging.FabricLogger, lastConfigBlock *common.Block) *assembler.Assembler {
-	assemblerGRPC := node2.CreateGRPCAssembler(conf)
+	assemblerGRPC := node_utils.CreateGRPCAssembler(conf)
 	assembler := assembler.NewAssembler(conf, assemblerGRPC, lastConfigBlock, logger)
 
 	orderer.RegisterAtomicBroadcastServer(assemblerGRPC.Server(), assembler)
